+++ src/main/java/org/jenkinsci/plugins/ghprb/GhprbTrigger.java
@@ -6,6 +6,14 @@ import com.google.common.annotations.VisibleForTesting;
 import hudson.Extension;
 import hudson.Util;
 import hudson.matrix.MatrixProject;
+import hudson.model.AbstractBuild;
+import hudson.model.AbstractProject;
+import hudson.model.Item;
+import hudson.model.ParameterDefinition;
+import hudson.model.ParameterValue;
+import hudson.model.ParametersDefinitionProperty;
+import hudson.model.Saveable;
+import hudson.model.StringParameterValue;
 import hudson.model.*;
 import hudson.model.queue.QueueTaskFuture;
 import hudson.plugins.git.util.BuildData;
@@ -18,7 +26,6 @@ import jenkins.model.Jenkins;
 import jenkins.model.ParameterizedJobMixIn;
 import net.sf.json.JSONObject;
 import org.apache.commons.lang.StringUtils;
-import org.jenkinsci.Symbol;
 import org.jenkinsci.plugins.ghprb.extensions.*;
 import org.jenkinsci.plugins.ghprb.extensions.comments.GhprbBuildLog;
 import org.jenkinsci.plugins.ghprb.extensions.comments.GhprbBuildResultMessage;
@@ -66,9 +73,11 @@ public class GhprbTrigger extends GhprbTriggerBackwardsCompatible {
     private String gitHubAuthId;
     private String triggerPhrase;
     private String skipBuildPhrase;
+    private String blackListCommitAuthor;
     private String blackListLabels;
     private String whiteListLabels;
-    
+    private String includedRegions;
+    private String excludedRegions;
 
     private transient Ghprb helper;
     private transient GhprbRepository repository;
@@ -115,6 +124,7 @@ public class GhprbTrigger extends GhprbTriggerBackwardsCompatible {
             Boolean displayBuildErrorsOnDownstreamBuilds,
             String commentFilePath,
             String skipBuildPhrase,
+            String blackListCommitAuthor,
             List<GhprbBranch> whiteListTargetBranches,
             List<GhprbBranch> blackListTargetBranches,
             Boolean allowMembersOfWhitelistedOrgsAsAdmin, 
@@ -125,7 +135,9 @@ public class GhprbTrigger extends GhprbTriggerBackwardsCompatible {
             String buildDescTemplate,
             String blackListLabels,
             String whiteListLabels,
-            List<GhprbExtension> extensions
+            List<GhprbExtension> extensions,
+            String includedRegions,
+            String excludedRegions
             ) throws ANTLRException {
         super(cron);
         this.adminlist = adminlist;
@@ -139,6 +151,7 @@ public class GhprbTrigger extends GhprbTriggerBackwardsCompatible {
         this.autoCloseFailedPullRequests = autoCloseFailedPullRequests;
         this.displayBuildErrorsOnDownstreamBuilds = displayBuildErrorsOnDownstreamBuilds;
         this.skipBuildPhrase = skipBuildPhrase;
+        this.blackListCommitAuthor = blackListCommitAuthor;
         this.whiteListTargetBranches = whiteListTargetBranches;
         this.blackListTargetBranches = blackListTargetBranches;
         this.gitHubAuthId = gitHubAuthId;
@@ -146,6 +159,8 @@ public class GhprbTrigger extends GhprbTriggerBackwardsCompatible {
         this.buildDescTemplate = buildDescTemplate;
         this.blackListLabels = blackListLabels;
         this.whiteListLabels = whiteListLabels;
+        this.includedRegions = includedRegions;
+        this.excludedRegions = excludedRegions;
         setExtensions(extensions);
         configVersion = latestVersion;
     }
@@ -478,6 +493,14 @@ public class GhprbTrigger extends GhprbTriggerBackwardsCompatible {
         return skipBuildPhrase;
     }
 
+    public String getBlackListCommitAuthor() {
+        if (StringUtils.isEmpty(blackListCommitAuthor)) {
+            // if it's empty grab the global value
+            return getDescriptor().getBlackListCommitAuthor();
+        }
+        return blackListCommitAuthor;
+    }
+
     public String getBlackListLabels() {
         if (blackListLabels == null) {
             return "";
@@ -542,6 +565,20 @@ public class GhprbTrigger extends GhprbTriggerBackwardsCompatible {
 
     public List<GhprbBranch> getBlackListTargetBranches() {
         return normalizeTargetBranches(blackListTargetBranches);
+    }
+
+    public String getIncludedRegions() {
+        if (includedRegions == null) {
+            return "";
+        }
+        return includedRegions;
+    }
+
+    public String getExcludedRegions() {
+        if (excludedRegions == null) {
+            return "";
+        }
+        return excludedRegions;
     }
 
     @Override
@@ -643,6 +680,7 @@ public class GhprbTrigger extends GhprbTriggerBackwardsCompatible {
         private String okToTestPhrase = ".*ok\\W+to\\W+test.*";
         private String retestPhrase = ".*test\\W+this\\W+please.*";
         private String skipBuildPhrase = ".*\\[skip\\W+ci\\].*";
+        private String blackListCommitAuthor = "";
         private String cron = "H/5 * * * *";
         private Boolean useComments = false;
         private Boolean useDetailedComments = false;
@@ -760,6 +798,7 @@ public class GhprbTrigger extends GhprbTriggerBackwardsCompatible {
             okToTestPhrase = formData.getString("okToTestPhrase");
             retestPhrase = formData.getString("retestPhrase");
             skipBuildPhrase = formData.getString("skipBuildPhrase");
+            blackListCommitAuthor = formData.getString("blackListCommitAuthor");
             cron = formData.getString("cron");
             useComments = formData.getBoolean("useComments");
             useDetailedComments = formData.getBoolean("useDetailedComments");
@@ -834,6 +873,10 @@ public class GhprbTrigger extends GhprbTriggerBackwardsCompatible {
 
         public String getSkipBuildPhrase() {
             return skipBuildPhrase;
+        }
+
+        public String getBlackListCommitAuthor() {
+            return blackListCommitAuthor;
         }
 
         public String getCron() {