+++ src/main/java/org/jenkinsci/plugins/ghprb/jobdsl/GhprbContextExtensionPoint.java
@@ -32,6 +32,7 @@ public class GhprbContextExtensionPoint extends ContextExtensionPoint {
                 context.displayBuildErrorsOnDownstreamBuilds,
                 null,
                 context.skipBuildPhrase,
+                context.blackListCommitAuthor,
                 context.whiteListTargetBranches,
                 context.blackListTargetBranches,
                 context.allowMembersOfWhitelistedOrgsAsAdmin,
@@ -42,7 +43,9 @@ public class GhprbContextExtensionPoint extends ContextExtensionPoint {
                 context.buildDescriptionTemplate,
                 Joiner.on("\n").join(context.blackListLabels),
                 Joiner.on("\n").join(context.whiteListLabels),
-                context.extensionContext.extensions
+                context.extensionContext.extensions,
+                context.includedRegions,
+                context.excludedRegions
         );
     }