+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ### Updates
 
+#### -> 1.38.0
+* Add ability to decide whether to build based on regions (or repository file paths as defined by regex)
+
+#### -> 1.37.0
+* Add user blacklist capability
+* Small but powerful docs cleanup
+
 #### -> 1.36.2
 * Revert pipeline support for the upstream status capability (only). See #506.